+++ wemake_python_styleguide/types.py
@@ -36,6 +36,13 @@ AnyImport = Union[ast.Import, ast.ImportFrom]
 #: In cases we need to work with both function definitions:
 AnyFunctionDef = Union[ast.FunctionDef, ast.AsyncFunctionDef]
 
+#: In cases we need to work with all function definitions (including Lambda)
+AnyFunctionDefAndLambda = Union[
+    ast.FunctionDef,
+    ast.AsyncFunctionDef,
+    ast.Lambda,
+]
+
 #: Flake8 API format to return error messages:
 CheckResult = Tuple[int, int, str, type]