+++ wemake_python_styleguide/violations/consistency.py
@@ -42,6 +42,8 @@ Summary
    ConstantComparisonViolation
    BadNumberSuffixViolation
    ComparisonOrderViolation
+   MultipleInComparisonViolation
+   RedundantComparisonViolation
 
 Consistency checks
 ------------------
@@ -57,6 +59,8 @@ Consistency checks
 .. autoclass:: ConstantComparisonViolation
 .. autoclass:: BadNumberSuffixViolation
 .. autoclass:: ComparisonOrderViolation
+.. autoclass:: MultipleInComparisonViolation
+.. autoclass:: RedundantComparisonViolation
 
 """
 
@@ -425,3 +429,71 @@ class BadNumberSuffixViolation(TokenizeViolation):
     #: Error message shown to the user.
     error_template = 'Found underscored number: {0}'
     code = 310
+
+
+class MultipleInComparisonViolation(ASTViolation):
+    """
+    Forbids comparision where multiple 'in's are userd in a statement.
+
+    Reasoning:
+        This is unreadable. Use different comparisons for it.
+        Bring a consistency to the comparison!
+
+    Solution:
+        Refactor your comparison expression to use several ``and`` conditions
+        or separate ``if`` statements in case it is appropriate.
+
+    Example::
+
+        # Correct:
+        if item in bucket and bucket in master_list_of_buckets:
+        if x_coord in line and line in square:
+
+        # Wrong:
+        if item in bucket in master_list_of_buckets:
+        if x_cord in line in square:
+
+    Note:
+        Returns Z311 as error code
+
+    """
+
+    should_use_text = False
+    #: Error message shown to the user.
+    error_template = 'Found multiple in comparisons'
+    code = 311
+
+
+class RedundantComparisonViolation(ASTViolation):
+    """
+    Forbids to have comparisons between the same variable.
+
+    Reasoning:
+        When the same variables are compared it is typically an indication
+        of a mistake, since the Boolean value of the comparison will always be
+        the same.
+
+    Solution:
+        Remove the same variable comparison and any associated dead code.
+
+    Example::
+
+        # Wrong:
+        a = 1
+        if a < a:
+            do_something()
+        else:
+            do_something_else()
+
+        # Correct:
+        do_something()
+
+    Note:
+        Returns Z312 as error code
+
+    """
+
+    should_use_text = False
+    #: Error message shown to the user.
+    error_template = 'Found comparison between same variable'
+    code = 312