+++ wemake_python_styleguide/visitors/ast/comparisons.py
@@ -7,6 +7,8 @@ from wemake_python_styleguide.types import AnyNodes
 from wemake_python_styleguide.violations.consistency import (
     ComparisonOrderViolation,
     ConstantComparisonViolation,
+    MultipleInComparisonViolation,
+    RedundantComparisonViolation,
 )
 from wemake_python_styleguide.visitors.base import BaseNodeVisitor
 
@@ -123,3 +125,59 @@ class WrongOrderVisitor(BaseNodeVisitor):
         """
         self._check_ordering(node)
         self.generic_visit(node)
+
+
+# TODO(@sobolevn): refactor to be a single visitor
+class MultipleInVisitor(BaseNodeVisitor):
+    """Restricts comparision where multiple `in`s are used."""
+
+    def _has_multiple_in_comparisons(self, node: ast.Compare) -> bool:
+        count = 0
+        for op in node.ops:
+            if isinstance(op, ast.In):
+                count += 1
+        return count > 1
+
+    def _count_in_comparisons(self, node: ast.Compare) -> None:
+        if self._has_multiple_in_comparisons(node):
+            self.add_violation(MultipleInComparisonViolation(node))
+
+    def visit_Compare(self, node: ast.Compare) -> None:
+        """
+        Forbids comparisons including multiple 'in's in a statement.
+
+        Raise:
+            MultipleInComparisonViolation
+
+        """
+        self._count_in_comparisons(node)
+        self.generic_visit(node)
+
+
+class RedundantComparisonVisitor(BaseNodeVisitor):
+    """Restricts the comparison where always same result."""
+
+    def visit_Compare(self, node: ast.Compare) -> None:
+        """
+        Ensures that compares are not for same variable.
+
+        Raises:
+            RedundantComparisonViolation
+
+        """
+        self._check_redundant_compare(node)
+        self.generic_visit(node)
+
+    def _is_same_variable(self, left: ast.AST, right: ast.AST) -> bool:
+        if isinstance(left, ast.Name) and isinstance(right, ast.Name):
+            if left.id is right.id:
+                return True
+        return False
+
+    def _check_redundant_compare(self, node: ast.Compare) -> None:
+        last_variable = node.left
+        for next_variable in node.comparators:
+            if self._is_same_variable(last_variable, next_variable):
+                self.add_violation(RedundantComparisonViolation(node))
+                break
+            last_variable = next_variable